refactor(desktop): move settings domain modules to features/settings
settings, verbosity, config-export moved. all importers updated.
