+++ apps/desktop/src/components/NewSessionDialog.tsx
@@ -24,7 +24,7 @@ import type {
 import { PROVIDER_CAPABILITIES, getDefaultTurnModel } from '@kay-am/core';
 import { shortModel } from '../features/session/agent-row-format';
 import { PROVIDER_LABEL_LOWER } from '../features/providers/providers';
-import { settingBranchPrefix, DEFAULT_BRANCH_PREFIX } from '../settings';
+import { settingBranchPrefix, DEFAULT_BRANCH_PREFIX } from '../features/settings/settings';
 import { SESSION_FEATURES } from '../shared/lib/features';
 import { EMPTY_ARRAY, useAppStore } from '../store';
 import { PlannerWidget } from './PlannerWidget';
