+++ apps/desktop/src/components/OpenInEditorButton.tsx
@@ -3,7 +3,11 @@ import { ChevronDown, FolderOpen } from 'lucide-react';
 import { Button, Tooltip, cn, type ButtonSize, type ButtonVariant } from '@kay-am/ui';
 import { openInEditor } from '../shared/lib/editor';
 import { formatError } from '../shared/lib/errors';
-import { DEFAULT_EDITOR_BINARY, SETTING_DEFAULT_EDITOR, SETTING_EDITOR_BINARY } from '../settings';
+import {
+  DEFAULT_EDITOR_BINARY,
+  SETTING_DEFAULT_EDITOR,
+  SETTING_EDITOR_BINARY,
+} from '../features/settings/settings';
 import { useAppStore } from '../store';
 
 interface OpenInEditorButtonProps {
