+++ apps/desktop/src/components/SettingsDialog.tsx
@@ -6,7 +6,7 @@ import { BudgetRulesPanel } from './BudgetRulesPanel';
 import { GithubPanel } from '../features/github/components/Panel';
 import { ImportConfigDialog } from './ImportConfigDialog';
 import type { ConfigBundleImportResult } from '@kay-am/types';
-import { DEFAULT_EDITOR_BINARY, SETTING_EDITOR_BINARY } from '../settings';
+import { DEFAULT_EDITOR_BINARY, SETTING_EDITOR_BINARY } from '../features/settings/settings';
 import { SESSION_FEATURES } from '../shared/lib/features';
 import { formatError } from '../shared/lib/errors';
 import { useAppStore } from '../store';
