+++ apps/desktop/src/components/WorkflowNextStepCta.tsx
@@ -2,7 +2,7 @@ import { useMemo, useState } from 'react';
 import { ArrowRight, ClipboardList } from 'lucide-react';
 import { cn } from '@kay-am/ui';
 import type { Agent, Step, Workflow } from '@kay-am/types';
-import type { VerbosityLevel } from '../verbosity';
+import type { VerbosityLevel } from '../features/settings/verbosity';
 import {
   AGENT_KIND_DEFAULTS,
   AGENT_KIND_PALETTE,
