+++ apps/desktop/src/components/WorkspaceSettingsDialog.tsx
@@ -6,7 +6,7 @@ import { SkillsPanel } from './SkillsPanel';
 import { PhasesPanel } from './PhasesPanel';
 import { BulkSessionDeleteDialog } from './WorkspacesSidebar';
 import { formatError } from '../shared/lib/errors';
-import { DEFAULT_BRANCH_PREFIX, settingBranchPrefix } from '../settings';
+import { DEFAULT_BRANCH_PREFIX, settingBranchPrefix } from '../features/settings/settings';
 import { WORKSPACE_FEATURES } from '../shared/lib/features';
 import { useAppStore, useSessions } from '../store';
